requirements.txt
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1

telebot~=0.0.5
peewee~=3.16.0
requests~=2.28.2